third_party/WebKit/LayoutTests/fast/dom/shadow/reinsert-insertion-point.html

<!DOCTYPE html>

<script src="../../../resources/js-test.js"></script>

<div id="inside">Should be visible</div>

<script>
var inside = document.getElementById('inside');
var insideRoot = inside.createShadowRoot();

var content = insideRoot.appendChild(document.createElement('content'));

shouldBeTrue("inside.offsetHeight > 0");

content.remove();
insideRoot.appendChild(content);

shouldBeTrue("inside.offsetHeight > 0");
</script>
